Renomeando projeto já existente para Domain, projeto que ficará o dominio da aplicação. E criando o projeto vazio Infrastructure que a primeiro momento vamos usar para registrar dependencias no container
