+++ TheatricalPlayersRefactoringKata.Domain/Performance.cs
@@ -1,4 +1,4 @@
-namespace TheatricalPlayersRefactoringKata;
+namespace TheatricalPlayersRefactoringKata.Domain;
 
 public class Performance
 {
