+++ TheatricalPlayersRefactoringKata.Domain/Play.cs
@@ -1,4 +1,4 @@
-namespace TheatricalPlayersRefactoringKata;
+namespace TheatricalPlayersRefactoringKata.Domain;
 
 public class Play
 {
