+++ TheatricalPlayersRefactoringKata.Infrastructure/Class1.cs
@@ -1,0 +1,5 @@
+﻿namespace TheatricalPlayersRefactoringKata.Infrastructure;
+
+public class Class1
+{
+}
