+++ TheatricalPlayersRefactoringKata.Tests/StatementPrinterTests.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using ApprovalTests;
 using ApprovalTests.Reporters;
+using TheatricalPlayersRefactoringKata.Domain;
 using Xunit;
 
 namespace TheatricalPlayersRefactoringKata.Tests;
